+++ lib/claper_web/router.ex
@@ -198,6 +198,9 @@ defmodule ClaperWeb.Router do
       live "/oidc_providers/new", OidcProviderLive, :new
       live "/oidc_providers/:id/edit", OidcProviderLive, :edit
       live "/oidc_providers/:id", OidcProviderLive, :show
+
+      live "/audit_logs", AuditLogLive, :index
+      live "/audit_logs/:id", AuditLogLive, :show
     end
   end
 end